Add joysound titlecard

+++ src/common/videoDownloader.ts
@@ -262,7 +262,9 @@ export function downloadJoysoundData(
       "base64"
     );
 
-    fs.writeFileSync(telopFilename, telopBuffer);
+    if (!fs.existsSync(telopFilename)) {
+      fs.writeFileSync(telopFilename, telopBuffer);
+    }
 
     const ffmpeg = spawn(
       resourcePaths.ffmpeg,
